refactor: colocate domain-specific test fixtures with their tests
Move test scenarios and helpers that are only used in single test files
from the shared test helper to their respective test files:

- Move TestScenarios object and createPartialContactInfo to core.test.tsx
- Move expectZoomInfoValues function to zoominfo.test.tsx
- Keep BASE_ZOOM_INFO constants in helper since they're used by the
  createMockZoomInfoHistory factory

This improves test organization by keeping domain-specific fixtures
with the tests that use them, while maintaining truly shared utilities
in the helper file.
